HW3/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu124
numpy==2.2.5
Pillow==11.2.1
pycocotools==2.0.8
rich==14.0.0
tifffile==2025.3.30
torch==2.5.1+cu124
torchvision==0.20.1+cu124
